Chore(*): Organize Classes
