+++ src/main/java/net/refractored/bloodmoon/boss/ZombieIBoss.java
@@ -9,12 +9,15 @@
  */
 
 
-package net.refractored.bloodmoon;
+package net.refractored.bloodmoon.boss;
 
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.Random;
 
+import net.refractored.bloodmoon.Bloodmoon;
+import net.refractored.bloodmoon.BloodmoonActuator;
+import net.refractored.bloodmoon.readers.LocaleReader;
 import org.bukkit.Location;
 import org.bukkit.Particle;
 import org.bukkit.Sound;
