+++ src/main/java/net/refractored/bloodmoon/commands/BloodmoonStart.java
@@ -1,8 +1,7 @@
 package net.refractored.bloodmoon.commands;
 
-import com.sk89q.worldedit.extension.platform.Actor;
 import net.refractored.bloodmoon.Bloodmoon;
-import net.refractored.bloodmoon.LocaleReader;
+import net.refractored.bloodmoon.readers.LocaleReader;
 import net.refractored.bloodmoon.PeriodicNightCheck;
 import org.bukkit.World;
 import org.bukkit.entity.Player;
