+++ src/main/java/net/refractored/bloodmoon/readers/LocaleReader.java
@@ -1,5 +1,6 @@
-package net.refractored.bloodmoon;
+package net.refractored.bloodmoon.readers;
 
+import net.refractored.bloodmoon.Bloodmoon;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.World;
